Week_2/Day5/project1_translator/requirements.txt
Flask
psycopg2
python-dotenv
gTTS
flask-cors
psycopg2-binary
psycopg
translate

